build-sys: Add build dependencies to Retro-0.6.gir
RetroGtk-0.6.gir depends on the .la files it needs so that 'make' knows
the corresponding libraries should be built before trying to generate the
gir file. Retro-0.6.gir does not have such dependency information, which
broke 'make -j6' on my system as 'make' was trying to generate the .gir
file before libretro-gobject.la was built.