gl-display: Drop the 'pixbuf' property
Instead, let accessing the pixbuf via the set_pixbuf() and get_pixbuf()
methods as previously done.

This avoid notifying changes of the pixbuf and accidentally building the
pixbuf on during the notification chain when building the pixbuf on
demand from the pixdata as will be introduced  in the next commit.