build: Use the appropriate dependency object
Instead of using `link_with` for internal targets, we should create a
dependency object for the libmanette shared library.

Not only this is the idiomatic form for internal dependencies in Meson,
but it also allows us to transitively depend on the library
dependencies, and to specify the inclusion path for the library itself,
thus avoiding the add_project_arguments() call using the build root
directory. This allows us to use libmanette as a subproject in other
Meson-based projects.