retro-core: Rework state and memory API
When getting state or memory, don't return anything and instead write the
result into a file. Similarly, when setting them, read them from a file.

This will be important when the core runs in a separate process, as it will
allow to only send the filename through IPC instead of the whole savestate,
which can be fairly large for some cores.

Since the get/set terminology doesn't make much sense anymore, rename:
* retro_core_get_state() -> retro_core_save_state()
* retro_core_set_state() -> retro_core_load_state()
* retro_core_get_memory() -> retro_core_save_memory()
* retro_core_set_memory() -> retro_core_load_memory()

Throw errors for retro_core_load/save_memory() as well, not just for state
functions.

Use temporary files for tests, and similarly rename them.