flatpak: Don't diff the test output
The output is not guaranteed to be stable and has changed recently, causing
every test to fail. Hence don't try to diff it and instead just check exit
code instead and show the whole output if it's not 0. If it's 0, assume the
test passed. Get rid of expected files as they aren't used anymore.