Add RetroMainLoopSource
This is a custom GSource that supports more precise delays than
GTimeoutSource, as it counts time in microseconds instead of milliseconds.

Handle rounding error by inserting an extra microsecond when needed.

Take extra care not to miss the ready time by setting the corresponding
timeout in prepare(), and exit immediately if we've missed the time.

In case we missed it after all, set ready time based on the previous ready
time if it's available, not on current time.

While this still isn't enough to make the timing 100% consistent, as other
sources, as well as blocking code, can increase the delay between frames,
it's an improvement over rounding the delay between frames to milliseconds.