retro-pa-player: Bufferize samples on each frame
Currently RetroCore bufferizes individual samples into batches of 512 frames.
It was done as a workaround for blocking audio, but it can lead to a situation
where the buffer was almost filled during a frame, but not completely, leading
to the audio being delayed until the next frame, which leads to audio desync.

Instead, do itin RetroPaPlayer itself, buffering the audio during a frame and
submitting it all at once when RetroCore emits 'frame' signal.

This will allow to stop doing buffering in RetroCore.