Introduce RetroGLRenderer
This implements an OpenGL renderer. Use EGL to get a context, this should
work on both Wayland and X11, except maybe some old drivers that only
support GLX.

Use a rather hacky way to get epoxy to provide pointers to OpenGL
functions, as it uses epoxy_gl* names internally, but it's better than
listing them all manually.

Unfortunately, for now we have to flip the scene on CPU if
bottom_left_origin=true, later we'll need to pass the framebuffer as is
with a flag, and instead flip it on UI process side by just drawing the
rectangle upside down for RetroGLDisplay, and applying a scale
transformation for RetroCairoDisplay.