core: Make it a singleton
As each core is guaranteed to live alone in its runner process, we don't
need a core call stack anymore to handle multple cores living in the
same process despite the Libretro API not supporting it. Instead, it is
replaced by a singleton, as we still need to know which the RetroCore
instance we should call back to.