core: Fix a reference leak
In 4b418a187df02897b0f6570888bf2cb5571500fc, g_set_object() was used to
simplify some code. This is wrong because it takes a reference on the
newer object, which the older code didn't do because it didn't need to.
The extra reference wasn't removed, leading to a leak.

This reverts that piece of code to what it was before the erroneous code
cleanup.